+++ src/cli/commands/node/mod.rs
@@ -1,5 +1,3 @@
 pub mod command;
-mod default;
 mod json;
 mod output;
-mod table;